requirements.txt
flask
rq
redis
pygithub
anthropic
httpx